refactor ZirconCode's Agent to only run when ENABLE_INSECURE_AGENTS is enabled in .env
