+++ .env.example
@@ -86,6 +86,10 @@ AWS_SANDBOX=false
 # You should not allow this on a shared Huginn box because it is not secure.
 ALLOW_JSONPATH_EVAL=false
 
+# Enable this setting to allow insecure Agents like the ShellCommandAgent.  Only do this
+# when you trust everyone using your Huginn installation.
+ENABLE_INSECURE_AGENTS=false
+
 # Use Graphviz for generating diagrams instead of using Google Chart
 # Tools.  Specify a dot(1) command path built with SVG support
 # enabled.
